+++ packages/core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @d-dash/core
 
+## 0.3.0
+
+### Minor Changes
+
+- feat: add variable substitution for widget text/html/titles, dependent chained queries, schema scoping for tenant, and the new @d-dash/builder SDK package.
+- feat: Add gauge/bar/pie/heatmap type to echart adapter. Add variables to core
+
 ## 0.2.0
 
 ### Minor Changes